radiation wrapper as a Stepper
