+++ workflows/prognostic_c48_run/runtime/steppers/radiation/_stepper.py
@@ -1,5 +1,17 @@
 import dataclasses
-from typing import Optional, Mapping, Any, Literal, Tuple, Sequence, Hashable
+import subprocess
+import os
+from typing import (
+    Optional,
+    MutableMapping,
+    Mapping,
+    Any,
+    Literal,
+    Tuple,
+    Sequence,
+    Hashable,
+)
+from mpi4py import MPI
 import cftime
 from datetime import timedelta
 import numpy as np
@@ -10,8 +22,9 @@ from runtime.steppers.machine_learning import (
     MultiModelAdapter,
     predict,
 )
-from runtime.types import State
+from runtime.types import State, Diagnostics
 from runtime.derived_state import DerivedFV3State
+from runtime.steppers.radiation._config import get_rad_config
 from radiation import RadiationDriver, getdata, NGPTSW, NGPTLW
 from vcm.calc.thermo.vertically_dependent import (
     pressure_at_interface,
@@ -23,98 +36,100 @@ P_REF = 1.0e5
 MINUTES_PER_HOUR = 60.0
 SECONDS_PER_MINUTE = 60.0
 
-DEFAULT_RAD_CONFIG = {
-    "imp_physics": 11,
-    "iemsflg": 1,  # surface emissivity control flag
-    "ioznflg": 7,  # ozone data source control flag
-    "ictmflg": 1,  # data ic time/date control flag
-    "isolar": 2,  # solar constant control flag
-    "ico2flg": 2,  # co2 data source control flag
-    "iaerflg": 111,  # volcanic aerosols
-    "ialbflg": 1,  # surface albedo control flag
-    "icldflg": 1,
-    "ivflip": 1,  # vertical index direction control flag
-    "iovrsw": 1,  # cloud overlapping control flag for sw
-    "iovrlw": 1,  # cloud overlapping control flag for lw
-    "isubcsw": 2,  # sub-column cloud approx flag in sw radiation
-    "isubclw": 2,  # sub-column cloud approx flag in lw radiation
-    "lcrick": False,  # control flag for eliminating CRICK
-    "lcnorm": False,  # control flag for in-cld condensate
-    "lnoprec": False,  # precip effect on radiation flag (ferrier microphysics)
-    "iswcliq": 1,  # optical property for liquid clouds for sw
-    "lsswr": True,
-    "lslwr": True,
-    "nfxr": 45,
-    "ncld": 5,
-    "ncnd": 5,
-    "lgfdlmprad": False,
-    "uni_cld": False,
-    "effr_in": False,
-    "indcld": -1,
-    "num_p3d": 1,
-    "npdf3d": 0,
-    "ncnvcld3d": 0,
-    "lmfdeep2": True,
-    "lmfshal": True,
-    "sup": 1.0,
-    "kdt": 1,
-    "do_sfcperts": False,
-    "pertalb": [[-999.0], [-999.0], [-999.0], [-999.0], [-999.0]],
-    "do_only_clearsky_rad": False,
-    "swhtr": True,
-    "solcon": 1320.8872136343873,
-    "lprnt": False,
-    "lwhtr": True,
-    "lssav": True,
-}
+LOOKUP_DATA_PATH = "gs://vcm-fv3gfs-serialized-regression-data/physics/lookupdata/lookup.tar.gz"  # noqa: E501
+FORCING_DATA_PATH = (
+    "gs://vcm-fv3gfs-serialized-regression-data/physics/forcing/*"  # noqa: 501
+)
 
 
 @dataclasses.dataclass
-class RadiationWrapperConfig:
+class RadiationConfig:
     """"""
 
     kind: Literal["python"]
     input_model: Optional[MachineLearningConfig] = None
 
 
-class RadiationWrapper:
+class RadiationStepper:
     def __init__(
-        self, input_model: Optional[MultiModelAdapter], driver: RadiationDriver,
+        self,
+        driver: RadiationDriver,
+        rad_config: MutableMapping[Hashable, Any],
+        comm: MPI.COMM_WORLD,
+        input_model: Optional[MultiModelAdapter],
     ):
-        self._input_model: Optional[MultiModelAdapter] = input_model
         self._driver: RadiationDriver = driver
-        self._rad_config: Optional[Mapping[str, Any]] = None
+        self._rad_config: MutableMapping[Hashable, Any] = rad_config
+        self._comm: MPI.COMM_WORLD = comm
+        self._input_model: Optional[MultiModelAdapter] = input_model
+        self._download_radiation_assets()
+        self._init_driver()
 
     @classmethod
-    def from_config(cls, config: RadiationWrapperConfig) -> "RadiationWrapper":
+    def from_config(
+        cls,
+        config: RadiationConfig,
+        comm: MPI.COMM_WORLD,
+        physics_namelist: Mapping[Hashable, Any],
+    ) -> "RadiationStepper":
+        rad_config = get_rad_config(physics_namelist)
         if config.input_model:
             model: Optional[MultiModelAdapter] = open_model(config.input_model)
         else:
             model = None
-        driver = RadiationDriver()
-        return cls(model, driver)
+        return cls(RadiationDriver(), rad_config, comm, model)
 
-    def rad_init(
+    def _download_radiation_assets(
         self,
-        rank: int,
-        physics_namelist: dict,
-        forcing_dir: str = "./data/forcing",
-        fv_core_dir: str = "./INPUT/",
-        default_rad_config: dict = DEFAULT_RAD_CONFIG,
+        lookup_data_path: str = LOOKUP_DATA_PATH,
+        forcing_data_path: str = FORCING_DATA_PATH,
+        lookup_local_dir: str = "./rad_data/lookup",
+        forcing_local_dir: str = "./rad_data/forcing",
     ) -> None:
+        """Gets lookup tables and forcing needed for the radiation scheme. TODO: Make lookup
+        data part of writing a run directory; make scheme able to read existing forcing.
+        """
+        if self._comm.rank == 0:
+            os.makedirs(lookup_local_dir)
+            copy_cmd = ["gsutil", "cp", lookup_data_path, lookup_local_dir]
+            subprocess.check_call(copy_cmd)
+            unpack_cmd = [
+                "tar",
+                "-xzvf",
+                os.path.join(lookup_local_dir, "lookup.tar.gz"),
+                "-C",
+                lookup_local_dir,
+            ]
+            subprocess.check_call(unpack_cmd)
+            os.makedirs(forcing_local_dir)
+            copy_cmd = ["gsutil", "cp", forcing_data_path, forcing_local_dir]
+            subprocess.check_call(copy_cmd)
+            unpack_cmd = [
+                "tar",
+                "-xzvf",
+                os.path.join(forcing_local_dir, "data.tar.gz"),
+                "-C",
+                forcing_local_dir,
+            ]
+            subprocess.check_call(unpack_cmd)
+        self._comm.barrier()
+        self._lookup_local_dir = lookup_local_dir
+        self._forcing_local_dir = forcing_local_dir
+
+    def _init_driver(self, fv_core_dir: str = "./INPUT/"):
         """Initialize the radiation driver"""
-        self._rad_config = self._get_rad_config(physics_namelist, default_rad_config)
         sigma = getdata.sigma(fv_core_dir)
         nlay = len(sigma) - 1
-        aerosol_data = getdata.aerosol(forcing_dir)
-        sfc_filename, sfc_data = getdata.sfc(forcing_dir)
-        solar_filename, _ = getdata.astronomy(forcing_dir, self._rad_config["isolar"])
-
+        aerosol_data = getdata.aerosol(self._forcing_local_dir)
+        sfc_filename, sfc_data = getdata.sfc(self._forcing_local_dir)
+        solar_filename, _ = getdata.astronomy(
+            self._forcing_local_dir, self._rad_config["isolar"]
+        )
         self._driver.radinit(
             sigma,
             nlay,
             self._rad_config["imp_physics"],
-            rank,
+            self._comm.rank,
             self._rad_config["iemsflg"],
             self._rad_config["ioznflg"],
             self._rad_config["ictmflg"],
@@ -138,72 +153,20 @@ class RadiationWrapper:
             sfc_data,
         )
 
-    @staticmethod
-    def _get_rad_config(physics_namelist, default_rad_config) -> Mapping[str, Any]:
-        """Generate radiation config from fv3gfs namelist to ensure
-        identical. Additional values from hardcoded defaults.
-        """
-        rad_config = {}
-        rad_config["imp_physics"] = physics_namelist["imp_physics"]
-        rad_config["iemsflg"] = physics_namelist["iems"]
-        rad_config["ioznflg"] = default_rad_config["ioznflg"]
-        rad_config["ictmflg"] = default_rad_config["ictmflg"]
-        rad_config["isolar"] = physics_namelist["isol"]
-        rad_config["ico2flg"] = physics_namelist["ico2"]
-        rad_config["iaerflg"] = physics_namelist["iaer"]
-        rad_config["ialbflg"] = physics_namelist["ialb"]
-        rad_config["icldflg"] = default_rad_config["icldflg"]
-        rad_config["ivflip"] = default_rad_config["ivflip"]
-        rad_config["iovrsw"] = default_rad_config["iovrsw"]
-        rad_config["iovrlw"] = default_rad_config["iovrlw"]
-        rad_config["isubcsw"] = physics_namelist["isubc_sw"]
-        rad_config["isubclw"] = physics_namelist["isubc_lw"]
-        rad_config["lcrick"] = default_rad_config["lcrick"]
-        rad_config["lcnorm"] = default_rad_config["lcnorm"]
-        rad_config["lnoprec"] = default_rad_config["lnoprec"]
-        rad_config["iswcliq"] = default_rad_config["iswcliq"]
-        rad_config["fhswr"] = physics_namelist["fhswr"]
-        rad_config["lsswr"] = default_rad_config["lsswr"]
-        rad_config["lslwr"] = default_rad_config["lslwr"]
-        rad_config["nfxr"] = default_rad_config["nfxr"]
-        rad_config["ncld"] = physics_namelist["ncld"]
-        rad_config["ncnd"] = physics_namelist["ncld"]
-        rad_config["fhswr"] = physics_namelist["fhswr"]
-        rad_config["fhlwr"] = physics_namelist["fhlwr"]
-        rad_config["lgfdlmprad"] = default_rad_config["lgfdlmprad"]
-        rad_config["uni_cld"] = default_rad_config["uni_cld"]
-        rad_config["effr_in"] = default_rad_config["effr_in"]
-        rad_config["indcld"] = default_rad_config["indcld"]
-        rad_config["num_p3d"] = default_rad_config["num_p3d"]
-        rad_config["npdf3d"] = default_rad_config["npdf3d"]
-        rad_config["ncnvcld3d"] = default_rad_config["ncnvcld3d"]
-        rad_config["lmfdeep2"] = default_rad_config["lmfdeep2"]
-        rad_config["lmfshal"] = default_rad_config["lmfshal"]
-        rad_config["sup"] = default_rad_config["sup"]
-        rad_config["kdt"] = default_rad_config["kdt"]
-        rad_config["do_sfcperts"] = default_rad_config["do_sfcperts"]
-        rad_config["pertalb"] = default_rad_config["pertalb"]
-        rad_config["do_only_clearsky_rad"] = default_rad_config["do_only_clearsky_rad"]
-        rad_config["swhtr"] = physics_namelist["swhtr"]
-        rad_config["solcon"] = default_rad_config["solcon"]
-        rad_config["lprnt"] = default_rad_config["lprnt"]
-        rad_config["lwhtr"] = physics_namelist["lwhtr"]
-        rad_config["lssav"] = default_rad_config["lssav"]
-
-        return rad_config
-
-    def rad_update(
+    def __call__(
         self,
+        state: State,
+        tracer_metadata: Mapping[str, Any],
         time: cftime.DatetimeJulian,
         dt_atmos: float,
-        forcing_dir: str = "./data/forcing",
-    ) -> None:
+    ):
+
+        self._rad_update(time, dt_atmos)
+        diagnostics = self._rad_compute(state, tracer_metadata, time, dt_atmos)
+        return {}, diagnostics, {}
+
+    def _rad_update(self, time: cftime.DatetimeJulian, dt_atmos: float) -> None:
         """Update the radiation driver's time-varying parameters"""
-        if self._rad_config is None:
-            raise ValueError(
-                "Radiation driver not initialized. `.rad_init` must be called "
-                "before `.rad_update`."
-            )
         # idat is supposed to be model initalization time but is unused w/ current flags
         idat = np.array(
             [time.year, time.month, time.day, 0, time.hour, time.minute, time.second, 0]
@@ -213,9 +176,11 @@ class RadiationWrapper:
         )
         fhswr = np.array(float(self._rad_config["fhswr"]))
         dt_atmos = np.array(float(dt_atmos))
-        aerosol_data = getdata.aerosol(forcing_dir)
-        _, solar_data = getdata.astronomy(forcing_dir, self._rad_config["isolar"])
-        gas_data = getdata.gases(forcing_dir, self._rad_config["ictmflg"])
+        aerosol_data = getdata.aerosol(self._forcing_local_dir)
+        _, solar_data = getdata.astronomy(
+            self._forcing_local_dir, self._rad_config["isolar"]
+        )
+        gas_data = getdata.gases(self._forcing_local_dir, self._rad_config["ictmflg"])
         self._driver.radupdate(
             idat,
             jdat,
@@ -231,20 +196,14 @@ class RadiationWrapper:
             gas_data,
         )
 
-    def rad_compute(
+    def _rad_compute(
         self,
         state: State,
         tracer_metadata: Mapping[str, Any],
         time: cftime.DatetimeJulian,
         dt_atmos: float,
-        rank: int,
-        lookup_dir: str = "./data/lookup",
-    ):
-        if self._rad_config is None:
-            raise ValueError(
-                "Radiation driver not initialized. `.rad_init` must be called "
-                "before `.rad_compute`."
-            )
+    ) -> Diagnostics:
+        """Compute the radiative fluxes"""
         if self._input_model is not None:
             predictions = predict(self._input_model, state)
             state = UpdatedState(state, predictions)
@@ -252,10 +211,12 @@ class RadiationWrapper:
         grid, coords = _grid(state)
         sfcprop = _sfcprop(state)
         ncolumns, nz = statein["tgrs"].shape[0], statein["tgrs"].shape[1]
-        model = _model(self._rad_config, tracer_metadata, time, dt_atmos, nz, rank)
+        model = _model(
+            self._rad_config, tracer_metadata, time, dt_atmos, nz, self._comm.rank
+        )
         random_numbers = getdata.random(ncolumns, nz, NGPTSW, NGPTLW)
-        lw_lookup = getdata.lw(lookup_dir)
-        sw_lookup = getdata.sw(lookup_dir)
+        lw_lookup = getdata.lw(self._lookup_local_dir)
+        sw_lookup = getdata.sw(self._lookup_local_dir)
         out = self._driver._GFS_radiation_driver(
             model, statein, sfcprop, grid, random_numbers, lw_lookup, sw_lookup
         )
@@ -286,7 +247,7 @@ TRACER_NAME_MAPPING = {  # this is specific to GFS physics
 
 
 def _model(
-    rad_config: Mapping[str, Any],
+    rad_config: MutableMapping[Hashable, Any],
     tracer_metadata: Mapping[str, Any],
     time: cftime.DatetimeJulian,
     dt_atmos: float,
@@ -302,6 +263,7 @@ def _model(
         "ncnd": rad_config["ncnd"],
         "fhswr": rad_config["fhswr"],
         "fhlwr": rad_config["fhlwr"],
+        # todo: why does solar hour need to be one timestep behind time to validate?
         "solhr": _solar_hour(time - timedelta(seconds=dt_atmos)),
         "lsswr": rad_config["lsswr"],
         "lslwr": rad_config["lslwr"],
@@ -394,8 +356,8 @@ def _unstack(
     coords: xr.DataArray,
     sample_dim: str = "column",
     unstacked_dim: str = "z",
-) -> Mapping[str, xr.DataArray]:
-    out = {}
+) -> Diagnostics:
+    out: Diagnostics = {}
     for name, arr in data.items():
         if arr.ndim == 1:
             da = xr.DataArray(arr, dims=[sample_dim], coords={sample_dim: coords})
